python-worker/requirements.txt
python-socketio>=5.3.0
aiohttp>=3.7.4